Mass renaming. Re-work the models. Initial fixes and at least one test passing (many many fail!)
- Massive renaming of value, items
- Re-worked the model now from treebeard
- Spent time debugging some of the first tests, to get some data into the new model.

Uncovering some issues with how the treebeard works verus the old. In the old, you could lookup all comments “for a type”. I can still do that here, but it won’t be a single queryset, so I’ll have to look at how that might — or could work. It may require a ‘Union’