Remove remaining dependencies on django_comments
After migrating all of the code that needed to be modified to work with this module, I found there was practically nothing left being used from django_comments. The only things were some templatetags, and some templates.

I migrated these to the tree module, and the django_comments is no longer needed.