Add assoc value to comments. Improve model relationships
I added a de-normalized ‘assoc’ value to each comment which is set to the “CommentAssociation” object connected to the ‘root’. This de-normalization ensables some significant improvements to the model queries, reducing the number of queries from 2 or 3 to 1 in most cases.

Updated the “CommentAssociation” ForeignKey to a OneToOne so that it is easier to reverse the relationship and use this in queries.

Solve an issue with the latest draftjs-exporter related to a block definition that is not correct for latest version.