Add user flag summary method. Expand tests
Added a method to get a summary of comments has flagged for a page. This method can be used by an API call to allow the UI to indicate which comments on a page a user has flagged as liked, disliked, etc.

Uses postgres ArrayAgg if postgres is use, or a local method to calulate the values if not.

Started expanding the tests so that I could perform tests on postgres as well as on sqllite or mysql — eventually